Coding 14

+++ controllers/deviceGroup_del_data.php
@@ -1,17 +1,38 @@
 <?php
+$data = array();
 
 include("../helpers/mysqli_connect.php");
 
 if(!empty($_POST['id'])){
     try {
+        //check relation data
+        $sql = "SELECT COUNT(device_group_id) as cnt FROM devices WHERE device_group_id = {$_POST['id']} ";
+        $result = $conn->query($sql);
+        if ($result->num_rows > 0) {
+            //output data of each row
+            while($row = $result->fetch_assoc()) {
+                $relate_cnt = $row['cnt'];
+            }
+        }
+        if ($relate_cnt > 0) {
+            $data[] = array("result" => false,
+                            "error" => 'Relation data, Can not delete');
+            echo json_encode($data);
+            return;
+        }
+
         //del data from the database
         $sql = "DELETE FROM device_group WHERE device_group_id = {$_POST['id']} ";
         $conn->query($sql);
 
-        echo json_encode(true);
+        $data[] = array("result" => true,
+                        "error" => '');
+        echo json_encode($data);
 
     } catch (Exception $e) {
-        echo json_encode(false);
+        $data[] = array("result" => false,
+                        "error" => $e);
+        echo json_encode($data);
     }
 }
